+++ src/main/java/mate/academy/model/Goods.java
@@ -1,5 +1,6 @@
 package mate.academy.model;
 
+import java.math.BigDecimal;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -21,5 +22,5 @@ public class Goods {
             allocationSize = 1)
     private Long id;
     private String name;
-    private Double cost;
+    private BigDecimal cost;
 }
